% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_BILLFISH}
\alias{SC_BILLFISH}
\title{Constant for the billfish species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_BILLFISH
}
\description{
Constant for the billfish species category
}
\keyword{datasets}
