% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_CETACEANS}
\alias{SC_CETACEANS}
\title{Constant for the cetaceans species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_CETACEANS
}
\description{
Constant for the cetaceans species category
}
\keyword{datasets}
