% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_NERITIC}
\alias{SC_NERITIC}
\title{Constant for the neritic tunas species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_NERITIC
}
\description{
Constant for the neritic tunas species category
}
\keyword{datasets}
