% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_OTHERS}
\alias{SC_OTHERS}
\title{Constant for the 'others' species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_OTHERS
}
\description{
Constant for the 'others' species category
}
\keyword{datasets}
