% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_RAYS}
\alias{SC_RAYS}
\title{Constant for the ray species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_RAYS
}
\description{
Constant for the ray species category
}
\keyword{datasets}
