% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_SEABIRDS}
\alias{SC_SEABIRDS}
\title{Constant for the seabirds species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_SEABIRDS
}
\description{
Constant for the seabirds species category
}
\keyword{datasets}
