% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_SEERFISH}
\alias{SC_SEERFISH}
\title{Constant for the seerfish species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_SEERFISH
}
\description{
Constant for the seerfish species category
}
\keyword{datasets}
