% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_SHARKS}
\alias{SC_SHARKS}
\title{Constant for the sharks species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_SHARKS
}
\description{
Constant for the sharks species category
}
\keyword{datasets}
