% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_TEMPERATE}
\alias{SC_TEMPERATE}
\title{Constant for the temperate tunas species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_TEMPERATE
}
\description{
Constant for the temperate tunas species category
}
\keyword{datasets}
