% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_TROPICAL}
\alias{SC_TROPICAL}
\title{Constant for the tropical tunas species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_TROPICAL
}
\description{
Constant for the tropical tunas species category
}
\keyword{datasets}
