% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_TUNAS_NEI}
\alias{SC_TUNAS_NEI}
\title{Constant for the tunas NEI species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_TUNAS_NEI
}
\description{
Constant for the tunas NEI species category
}
\keyword{datasets}
