% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sc.R
\docType{data}
\name{SC_TURTLES}
\alias{SC_TURTLES}
\title{Constant for the turtles species category}
\format{
An object of class \code{character} of length 1.
}
\usage{
SC_TURTLES
}
\description{
Constant for the turtles species category
}
\keyword{datasets}
