% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sg.R
\docType{data}
\name{SG_BILLFISH}
\alias{SG_BILLFISH}
\title{Constant for the billfish species group}
\format{
An object of class \code{character} of length 1.
}
\usage{
SG_BILLFISH
}
\description{
Constant for the billfish species group
}
\keyword{datasets}
