% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sg.R
\docType{data}
\name{SG_OTHERS}
\alias{SG_OTHERS}
\title{Constant for the 'others' species group}
\format{
An object of class \code{character} of length 1.
}
\usage{
SG_OTHERS
}
\description{
Constant for the 'others' species group
}
\keyword{datasets}
