% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sg.R
\docType{data}
\name{SG_SEERFISH}
\alias{SG_SEERFISH}
\title{Constant for the seerfish species group}
\format{
An object of class \code{character} of length 1.
}
\usage{
SG_SEERFISH
}
\description{
Constant for the seerfish species group
}
\keyword{datasets}
