% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sg.R
\docType{data}
\name{SG_SHARKS}
\alias{SG_SHARKS}
\title{Constant for the sharks species group}
\format{
An object of class \code{character} of length 1.
}
\usage{
SG_SHARKS
}
\description{
Constant for the sharks species group
}
\keyword{datasets}
