% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_sg.R
\docType{data}
\name{SG_TUNAS}
\alias{SG_TUNAS}
\title{Constant for the tunas species group}
\format{
An object of class \code{character} of length 1.
}
\usage{
SG_TUNAS
}
\description{
Constant for the tunas species group
}
\keyword{datasets}
