% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_wp.R
\docType{data}
\name{WP_BILL}
\alias{WP_BILL}
\title{Constant for the WP on billifsh}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_BILL
}
\description{
Constant for the WP on billifsh
}
\keyword{datasets}
