% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_wp.R
\docType{data}
\name{WP_BYCT}
\alias{WP_BYCT}
\title{Constant for the WP on ecosystems and bycatch}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_BYCT
}
\description{
Constant for the WP on ecosystems and bycatch
}
\keyword{datasets}
