% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_wp.R
\docType{data}
\name{WP_NERI}
\alias{WP_NERI}
\title{Constant for the WP on neritic tunas}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_NERI
}
\description{
Constant for the WP on neritic tunas
}
\keyword{datasets}
