% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_wp.R
\docType{data}
\name{WP_TEMP}
\alias{WP_TEMP}
\title{Constant for the WP on temperate tunas}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_TEMP
}
\description{
Constant for the WP on temperate tunas
}
\keyword{datasets}
