% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_wp.R
\docType{data}
\name{WP_TROP}
\alias{WP_TROP}
\title{Constant for the WP on tropical tunas}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_TROP
}
\description{
Constant for the WP on tropical tunas
}
\keyword{datasets}
