% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_wp.R
\docType{data}
\name{WP_UNCL}
\alias{WP_UNCL}
\title{Constant for a generic 'unclassified' WP}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_UNCL
}
\description{
Constant for a generic 'unclassified' WP
}
\keyword{datasets}
